Added multiple image support for Flutter Phone App
Each Profile has a list of images that represents the different pictures that are taken of the same character during the trial of the film.

The UI has been updated to support the switching of images.
State management now also has support for multiple images in a profile.

Some minor bugs have been fixed as well.

27/1/20