Added multiple profiles support for Flutter Phone App
The JSON parser has been adapted to work with multiple Profiles.
Buttons to switch Profiles have been added to the UI without transitions.

Error handling is yet to be fully completed, particularly in the cases where the JSON is an empty list and where the images list of a profile is empty.

28/1/20