Medical records page for projector
The page is meant to show the injuries of Bill Walters and Henry Lang (The Father).

It is static and rather rigid. It can do with a subcomponent instead of two HTML instances that were copy pasted.

Some fixes were also made to controller, but the occasional disconnect and weird behavior still persists though a reload still fixes it as well.

28/2/20